% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rate.R
\name{exchange_rate_decomposition}
\alias{exchange_rate_decomposition}
\title{Decompostion into exchange rate and non exchange rate effect}
\usage{
\method{get_diff_portfolio_value}{data.frame}(sp_exch_rate_pair,
  ap_start_date, ap_end_date, np_mthly_yearly, dp_dates_investment_value)
}
\arguments{
\item{sp_exch_rate_pair}{exchange rate pair. e.g "USDSGD=X". "<Foreign_currency><local_currency>=X"}

\item{ap_start_date}{starting date of portfolio e.g. 2017-10-01}

\item{ap_end_date}{ending date of portfolio e.g. 2020-10-01. If you include a date beyond current date, the function will use the current date instead}

\item{np_mthly_yearly}{#Decomposition at monthly or yearly level}

\item{dp_dates_investment_value}{#data frame of date and investment values}
}
\description{
This function allows you to do strip out exchange effects (e.g. investment portfolio, revenue)
Idea is to first reverse all the SGD denominated value over time into foreign currency value. Then apply a constant exchange rate at beginning of the period over time
}
\keyword{exchangeRate}
