% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rate.R
\name{exchRate_effect_decomposition}
\alias{exchRate_effect_decomposition}
\title{Decompostion into exchange rate and non exchange rate effect}
\usage{
exchRate_effect_decomposition(sp_exch_rate_pair, ap_start_date,
  ap_end_date, np_mthly_yearly, dp_dates_investment_value)
}
\arguments{
\item{sp_exch_rate_pair}{#exchange rate pair}

\item{ap_start_date}{#starting date of portfolio}

\item{ap_end_date}{#ending date of portfolio (include a default here)}

\item{np_mthly_yearly}{#Decomposition at monthly or yearly level}

\item{dp_dates_investment_value}{#data frame of date and investment values}
}
\value{
object
}
\description{
This function allows you to do strip out exchange effects (e.g. investment portfolio, revenue)
Idea is to first reverse all the SGD denominated value over time into foreign currency value. Then apply a constant exchange rate at beginning of the period over time
Need to find a way to merge dp_dates_investment_value into the frequency of dates (through month or year)
}
