% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exchange_rate.R
\name{multiple_exchange_rate_decomposition}
\alias{multiple_exchange_rate_decomposition}
\title{This function allows you to do strip out exchange effects (e.g. investment portfolio, revenue) from multiple instruments at a time.}
\usage{

  \method{multiple_index_exchange_rate_decomposition}{data.frame}(np_number_of_instruments)
}
\arguments{
\item{np_number_of_instruments}{number of instruments to carry out the exchange rate decomposition}
}
\description{
This function allows you to do strip out exchange effects (e.g. investment portfolio, revenue) from multiple instruments at a time.
}
\keyword{exchangeRate}
